docker/flask/requirements.txt
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
DateTime==4.7
decorator==5.1.1
Flask==2.0.3
idna==3.4
importlib-metadata==4.8.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
MarkupSafe==2.0.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
ptyprocess==0.7.0
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.27.1
six==1.16.0
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.0.3
zipp==3.6.0
zope.interface==5.5.2
gunicorn==20.0.4
